+++ app/containers/HomePage.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component } from 'react';
-import Home from '../components/Home';
+import Home from '../features/home/Home';
 
 type Props = {};
