+++ app/containers/Root.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'connected-react-router';
 import { hot } from 'react-hot-loader/root';
-import type { Store } from '../reducers/types';
+import type { Store } from '../types';
 import Routes from '../Routes';
 
 type Props = {
