+++ app/features/home/Home.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import routes from '../constants/routes.json';
+import routes from '../../constants/routes.json';
 import styles from './Home.css';
 
 type Props = {};
